refactor(login): modulariza a página de login
