+++ frontend/src/components/form/FormField.tsx
@@ -1,17 +1,13 @@
-import React from 'react';
+import React from "react";
 
 interface FormFieldProps {
   label: string;
-  type?: string;
-  placeholder?: string;
-  value?: string;
-  onChange?: (value: string) => void;
-  onBlur?: () => void;
+  type: string;
+  placeholder: string;
+  value: string;
+  onChange: (value: string) => void;
   error?: string;
-  readOnly?: boolean;
   required?: boolean;
-  className?: string;
-  children?: React.ReactNode;
 }
 
 export const FormField: React.FC<FormFieldProps> = ({
@@ -20,32 +16,31 @@ export const FormField: React.FC<FormFieldProps> = ({
   placeholder,
   value,
   onChange,
-  onBlur,
   error,
-  readOnly = false,
   required = false,
-  className = "",
-  children,
 }) => {
   return (
-    <div className={`space-y-2 ${className}`}>
-      <label className="block text-sm font-semibold text-accent-700">
+    <div className="space-y-2">
+      <label className="block text-sm font-medium text-gray-700">
         {label}
         {required && <span className="text-red-500 ml-1">*</span>}
       </label>
-      {children || (
-        <input
-          type={type}
-          placeholder={placeholder}
-          value={value}
-          onChange={(e) => onChange?.(e.target.value)}
-          onBlur={onBlur}
-          className={`input-field ${error ? 'input-error' : ''}`}
-          readOnly={readOnly}
-        />
-      )}
+      <input
+        type={type}
+        placeholder={placeholder}
+        value={value}
+        onChange={(e) => onChange(e.target.value)}
+        className={`w-full px-4 py-3 border rounded-lg focus:ring-2 focus:ring-primary-500 focus:border-transparent transition-colors ${
+          error
+            ? "border-red-500 focus:ring-red-500"
+            : "border-gray-300"
+        }`}
+        required={required}
+      />
       {error && (
-        <span className="error-message">{error}</span>
+        <div className="ml-2 mt-1">
+          <p className="text-red-600 text-sm">{error}</p>
+        </div>
       )}
     </div>
   );
